Add login_required and is_superuser check

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 from .models import Product, Category
@@ -122,8 +123,13 @@ def all_specials(request):
     return render(request, 'products/all_specials.html', context)
 
 
+@login_required
 def add_product(request):
     """ Add a product to the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owner can add products.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -131,7 +137,9 @@ def add_product(request):
             messages.success(request, "Product was added successfully.")
             return redirect(reverse('product_detail', args=[product.id]))
         else:
-            messages.error(request, 'Product was not added. Correct the form inputs.')
+            messages.error(request,
+                            'Product was not added. Correct the form inputs.'
+                            )
     else:
         form = ProductForm()
 
@@ -143,8 +151,13 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """ Edit a product in the bookstore offer """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owner can edit products.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -153,7 +166,9 @@ def edit_product(request, product_id):
             messages.success(request, 'Product successfully updated!')
             return redirect(reverse('product_detail', args=[product.id]))
         else:
-            messages.error(request, 'Failed to update product. Please check the form inputs.')
+            messages.error(request,
+                            'Failed to update product. Please check the form inputs.'
+                            )
     else:
         form = ProductForm(instance=product)
         messages.info(request, f'You are editing {product.title}')
@@ -167,8 +182,13 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """ Delete a product from the store offer """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owner can delete products.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product was deleted.')
